Fix for TagLibToDirectiveAndFunction to respect semantics of the body() closure.
Currently, if body() is executed in a Grails taglib called from a Freemarker
template, TagLibToDirectiveAndFunction will simply spit the body() out into
output stream and return an empty string to the taglib. This works for *some*
use cases where the body of the taglib call is just pasted into output, but for
transformative stuff this doesn't work. For example if we have a taglib that
reverses the body and outputs it, that kind of functionality wouldn't work
without this patch.